// src/types/TimerActionType.ts
export enum TimerActionType {
  ADD_TIMER = 'ADD_TIMER',
  UPDATE_TIMER = 'UPDATE_TIMER',
  DELETE_TIMER = 'DELETE_TIMER',
  START_TIMER = 'START_TIMER',
  PAUSE_TIMER = 'PAUSE_TIMER',
  RESET_TIMER = 'RESET_TIMER',
  COMPLETE_TIMER = 'COMPLETE_TIMER',
  START_CATEGORY = 'START_CATEGORY',
  PAUSE_CATEGORY = 'PAUSE_CATEGORY',
  RESET_CATEGORY = 'RESET_CATEGORY',
  ADD_TO_HISTORY = 'ADD_TO_HISTORY',
  INIT_STATE = 'INIT_STATE'
}